fix(dada): widen nalign/nshroud to u64 — they wrap on large pools
`b.nalign` and `b.nshroud` were `u32` and are incremented once per
comparison. On the NovaSeq soil 16S pool they overflow twice, and the
archived #139 logs show it exactly:

    screen comps      11,605,515,622   (u64, correct)
    ALIGN: aligns      3,015,581,030   = 11,605,515,622 - 2*2^32
    ALIGN: shrouded    2,493,173,397   = 11,083,107,989 - 2*2^32

The wrap is not confined to the verbose `ALIGN:` line: both counters
reach `DadaResult` and are serialised into the output JSON, so any pool
above ~4.3e9 comparisons has been writing wrong values into user-facing
artifacts. Release builds wrap silently; debug builds would panic on the
same input.

Widened to u64 through `B`, `DadaResult`, and the two serialisation
structs. No behaviour depends on either counter, so clustering output is
unchanged.

Found while mining the #139 archives for #142. The #142 screen figures
themselves are unaffected — they are computed from `n_cmp_screened` and
`n_cmp_aligned`, which were already u64.

+++ src/containers.rs
@@ -286,9 +286,9 @@ pub struct B {
     /// Total read count across all Raws.
     pub reads: u32,
     /// Number of pairwise alignments performed.
-    pub nalign: u32,
+    pub nalign: u64,
     /// Number of comparisons screened out by k-mer distance ("shrouded").
-    pub nshroud: u32,
+    pub nshroud: u64,
     /// Significance threshold for abundance-based cluster splitting.
     pub omega_a: f64,
     /// Significance threshold for singleton detection.
